styles for LibrariesViews components are moved to a separate file
