+++ src/LibrariesList/LibrariesMapView.tsx
@@ -1,11 +1,9 @@
 import React, {useRef, FC} from 'react';
-import {View, StyleSheet, Dimensions, PixelRatio, Platform} from 'react-native';
+import {View, PixelRatio, Platform} from 'react-native';
 import MapView, {Marker} from 'react-native-maps';
 import {Loader} from '../components';
 import {MapViewProps} from './models';
-
-const windowWidth = Dimensions.get('window').width;
-const windowHeight = Dimensions.get('window').height;
+import styles from './styles';
 
 const LibrariesMapView: FC<MapViewProps> = ({data, loading, location}) => {
   const mapRef = useRef<{current: typeof MapView}>();
@@ -52,13 +50,5 @@ const LibrariesMapView: FC<MapViewProps> = ({data, loading, location}) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  map: {
-    width: windowWidth,
-    height: windowHeight,
-    marginBottom: 1,
-  },
-});
 
 export default LibrariesMapView;
